extract header to reusable component
