+++ frontend/src/App.vue
@@ -1,22 +1,6 @@
 <template>
   <q-layout view="hHh lpR fFf">
-    <q-header elevated class="bg-green-7 text-white">
-      <q-toolbar>
-        <q-btn flat dense round @click="drawer = !drawer">
-          <transition name="rotate" mode="out-in">
-            <q-icon :key="drawer" :name="drawer ? 'close' : 'menu'" />
-          </transition>
-        </q-btn>
-        <q-toolbar-title class="text-center">FPApp</q-toolbar-title>
-        <div class="row items-center no-wrap q-gutter-sm">
-          <q-icon name="account_circle" />
-          <div class="column items-end">
-            <div>User: ryuji</div>
-            <div>{{ currentTime }}</div>
-          </div>
-        </div>
-      </q-toolbar>
-    </q-header>
+    <AppHeader v-model:drawer="drawer" />
 
     <q-drawer v-model="drawer" show-if-above bordered overlay>
       <q-list>
@@ -36,16 +20,17 @@
 </template>
 
 <script setup lang="ts">
-import { ref, computed, onMounted, onBeforeUnmount } from 'vue';
-import DashboardView from './components/DashboardView.vue';
-import DataEntryView from './components/DataEntryView.vue';
+import { ref, computed } from 'vue';
+import AppHeader from './components/AppHeader.vue';
+import Dashboard from './components/Dashboard.vue';
+import DataEntry from './components/DataEntry.vue';
 
 const drawer = ref(false);
 const currentView = ref<'dashboard' | 'data-entry'>('dashboard');
 
 const componentMap = {
-  dashboard: DashboardView,
-  'data-entry': DataEntryView
+  dashboard: Dashboard,
+  'data-entry': DataEntry
 } as const;
 
 const currentComponent = computed(() => componentMap[currentView.value]);
@@ -54,35 +39,5 @@ function switchView(view: 'dashboard' | 'data-entry') {
   currentView.value = view;
   drawer.value = false;
 }
-
-const currentTime = ref('');
-let timer: number;
-
-function updateTime() {
-  const now = new Date(Date.now() + 9 * 60 * 60 * 1000);
-  const y = now.getUTCFullYear();
-  const m = String(now.getUTCMonth() + 1).padStart(2, '0');
-  const d = String(now.getUTCDate()).padStart(2, '0');
-  const h = String(now.getUTCHours()).padStart(2, '0');
-  const min = String(now.getUTCMinutes()).padStart(2, '0');
-  currentTime.value = `${y}/${m}/${d} ${h}:${min}`;
-}
-
-onMounted(() => {
-  updateTime();
-  timer = window.setInterval(updateTime, 1000);
-});
-
-onBeforeUnmount(() => {
-  clearInterval(timer);
-});
 </script>
 
-<style scoped>
-.rotate-enter-active, .rotate-leave-active {
-  transition: transform 0.2s ease;
-}
-.rotate-enter-from, .rotate-leave-to {
-  transform: rotate(180deg);
-}
-</style>
