--- frontend/src/components/DashboardView.vue
@@ -1,23 +1,0 @@
-<template>
-  <q-page class="q-pa-md">
-    <q-card flat bordered class="q-pa-md">
-      <q-card-section>
-        <div class="text-h6">Mock Dashboard</div>
-      </q-card-section>
-      <q-table :rows="rows" :columns="columns" row-key="id" flat />
-    </q-card>
-  </q-page>
-</template>
-
-<script setup lang="ts">
-const columns = [
-  { name: 'item', label: 'Item', field: 'item', align: 'left' },
-  { name: 'amount', label: 'Amount', field: 'amount', align: 'right',
-    format: (val: number) => `$${val.toFixed(2)}` }
-];
-
-const rows = [
-  { id: 1, item: 'Income', amount: 5000 },
-  { id: 2, item: 'Expenses', amount: 3000 }
-];
-</script>
